### 2022-01-21 #4088 @whedon
[PRE REVIEW]: Imagedata: A Python library to handle medical image data in NumPy array subclass Series
**Submitting author:** @erling6232 (<a href="http://orcid.org/0000-0001-8268-1206">Erling Andersen</a>)
**Repository:** <a href="https://github.com/erling6232/imagedata" target ="_blank">https://github.com/erling6232/imagedata</a>
**Version:** 1.4.0
