+++ paper.bib
@@ -127,6 +127,7 @@ howpublished={Presented at the 10th Annual Meeting of the Organization for Human
   author = {S M Smith and M Jenkonson and M W Woolrich and C F Beckmann and T E J Behrens and
             H Johansen-Berg and P R Bannister and M De Luca and I Drobnjak and D E Flitney and
             R Niazy and J Saunders and J Vickers and Y Zhang and N De Stefano and J M Brady and P M Matthews},
+  DOI={10.1016/j.neuroimage.2004.07.051}
   journal = {Neuro{I}mage},
   year = {2004},
   volume = {23(S1)},
