### 2022-02-07 0781590
viewer.get_level() and .get_window(): Handle DICOM input where multiple window center/level values are given. Will use the first parameter pair.
