+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.1dev0
+1.4.1
