### 2022-02-25 dc21292
read(): Handle signed integer depending on value of Pixel Representation.
