### 2022-02-25 6309b38
write_slice(): Set Pixel Representation depending on dtype.
