### 2022-03-07 29a3235
Add test to write np.eye to file.
