### 2022-03-09 erling6232/imagedata#3 @mwegrzyn
JOSS Review
Dear @erling6232,

my name is Martin Wegrzyn and I am one of the reviewers assigned to your submission to JOSS. Thank you for your work on Imagedata. The following are some comments I have regarding my first impression of the package and the manuscript:

**installation**

- [x]  I had problems installing the software which I think were related to the itk package. For me, the installation only worked if I did (in a clean environment, like: ```conda create -n imdata python=3```)
1. ```conda install -c https://conda.anaconda.org/simpleitk SimpleITK```
2. ```python -m pip install imagedata```

Otherwise, I got error messages first regarding scikit-build missing, then regarding problems with cmake, then regarding itk - which I could not solve. If you are interested, I can share the full history of error messages in a separate comment. However, I had these problems on both Linux and Windows and therefore suspect that others might run into similar issues as well. Therefore, I think that specific troubleshooting information, e.g. regarding the itk package, could be helpful.

**naming**

- [x] there are already may repositories on github named "imagedata" and it is not a search engine friendly name. Do you think that the naming of the package is sufficiently specific?

**manuscript**

- [x] I found the manuscript well-written and helpful. The motivation is clear to me. I would like to know however, if the tool is recommended for a specific area of clinical work and/or in which area it has been developed (e.g. brain, spine, heart etc.)

**information regarding brain imaging**

- [x] I work in the area of neuroimaging and use mainly dcm2niix for transformation from the DICOM to the nifti format (https://github.com/neurolabusc/dcm2niix). I find "imagedata" very attractive, because I could make it part of my python workflow and I like the possibility to transform niftis back to DICOM. However, one main question in neuroimaging that I did not see addressed (sorry if I overlooked it!) is the question of orientation. I.e. where is left and right in the image - or in other words - does the software use radiological (left is right) or neurological (left is left) orientation?

- [x] related to the previous question: how can I make sure that left and right have not been inadvertently flipped, either going from or back to the DICOM format? It would be important that left/right information is not only available to "imagedata" itself but that other software packages (FSL, SPM, FreeSurfer, AFNI, nilearn) will interpret it in the same way. Overall, left-right flipping is a common fear that I think should be addressed (e.g. https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/Orientation%20Explained). Maybe there is an automated test that I overlooked. If so, I think it could be helpful to highlight it. If not, I think this issue should be adressed by explicit tests.

- [x] if I have a mosaic of a 3D-image, which is common in functional imaging (e.g. https://github.com/neurolabusc/dcm_qa_mosaic), how can I convert the slices into a 3D-image? could I convert a 3D-image back into a DICOM mosaic and get the same slice order as in the original mosaic?

**Documentation**

- [x] Community guidelines could be elaborated (how to contribute, how to get help)

I hope you find these comments to be helpful. Please let me know if you have any questions regarding the points I raised. Thank you for your time and consideration.

Best regards,
Martin Wegrzyn
