### 2022-03-09 #4133 @mwegrzyn
Dear all,

I have finished my review and opened an issue outlining my comments [here](https://github.com/erling6232/imagedata/issues/3#issue-1164224377).

Best,
Martin
