### 2022-03-15 fe980e7
Header.set_default_values(): Set default geometry.
