### 2022-03-15 97633e6
ITKPlugin.get_image_from_numpy(): return image as np.ndarray, not itk matrix.
