### 2022-03-16 #4133 @galessiorob
Thank you so much for your review, @mwegrzyn! Would you mind outlining what about the installation process is not covered in the documentation?

@erling6232 per @mwegrzyn 's review, could you please consider adding automated tests and community guidelines?
