+++ CONTRIBUTING.rst
@@ -1,0 +1,207 @@
+############
+CONTRIBUTING
+############
+
+This is the guide for contributing code, documentation and tests, and for
+filing issues. Please read it carefully to help make the code review
+process go as smoothly as possible and maximize the likelihood of your
+contribution being merged.
+
+_Note:_  
+If you want to contribute new functionality, you may first consider if this 
+functionality belongs to the imagedata core, or is better suited for
+a plugin or an application program.
+[contrib-pydicom](https://github.com/pydicom/contrib-pydicom). contrib-pydicom
+If you're not sure where your contribution belongs,
+create an issue where you can discuss this before creating a pull request.
+
+
+How to contribute
+-----------------
+
+The preferred workflow for contributing to imagedata is to fork the
+[main repository](https://github.com/erling6232/imagedata) on
+GitHub, clone, and develop on a branch. Steps:
+
+1. Fork the [project repository](https://github.com/erling6232/imagedata)
+   by clicking on the 'Fork' button near the top right of the page. This creates
+   a copy of the code under your GitHub user account. For more details on
+   how to fork a repository see [this guide](https://help.github.com/articles/fork-a-repo/).
+
+2. Clone your fork of the imagedata repo from your GitHub account to your local disk:
+
+.. code-block:: bash
+
+   $ git clone git@github.com:YourLogin/imagedata.git
+   $ cd imagedata
+
+3. Create a ``feature`` branch to hold your development changes:
+
+.. code-block:: bash
+
+   $ git checkout -b my-feature
+
+   Always use a ``feature`` branch. It's good practice to never work on the ``master`` branch!
+
+4. Develop the feature on your feature branch. Add changed files using ``git add`` and then ``git commit`` files:
+
+.. code-block:: bash
+
+   $ git add modified_files
+   $ git commit
+
+5. Add a meaningful commit message. Pull requests are "squash-merged", e.g.
+   squashed into one commit with all commit messages combined. The commit
+   messages can be edited during the merge, but it helps if they are clearly
+   and briefly showing what has been done in the commit. Check out the 
+   [seven commonly accepted rules](https://www.theserverside.com/video/Follow-these-git-commit-message-guidelines)
+   for commit messages.
+   
+6. To record your changes in Git, push the changes to your GitHub
+   account with:
+
+.. code-block:: bash
+
+   $ git push -u origin my-feature
+
+7. Follow [these instructions](https://help.github.com/articles/creating-a-pull-request-from-a-fork)
+   to create a pull request from your fork. This will send an email to the committers.
+
+(If any of the above seems like magic to you, please look up the
+[Git documentation](https://git-scm.com/documentation) on the web, or ask a friend or another contributor for help.)
+
+Pull Request Checklist
+----------------------
+
+We recommend that your contribution complies with the following rules before you
+submit a pull request:
+
+-  Follow the style used in the rest of the code. That mostly means to
+   follow [PEP-8 guidelines](https://www.python.org/dev/peps/pep-0008/) for
+   the code, and the [Google style](https://github.com/google/styleguide/blob/gh-pages/pyguide.md#38-comments-and-docstrings)
+   for documentation.
+   
+-  If your pull request addresses an issue, please use the pull request title to
+   describe the issue and mention the issue number in the pull request
+   description. This will make sure a link back to the original issue is
+   created. Use "closes #issue-number" or "fixes #issue-number" to let GitHub 
+   automatically close the related issue on commit. Use any other keyword 
+   (i.e. works on, related) to avoid GitHub to close the referenced issue.
+
+-  All public methods should have informative docstrings with sample
[…]
