### 2022-03-21 e57861e
image_calculator: Added usage examples where input is np.eye().
