### 2022-03-21 ba6d015
Series.__new__(): Set default header when input is np.ndarray.
