### 2022-03-21 6ede9f7
cmdline.add_argparse_options(): Default template and geometry option values are None.
