### 2022-03-21 0937e3f
image_data.calculator(): Accept no indirs arguments. image_data.calculator(): Accept template and geometry options when given. image_data.calculator(): When no indirs arguments, calculate a new image, optionally with template and geometry given.
