### 2022-03-21 f45a81c
DICOMPlugin.write_slice(): Refactor calculation of smallest/largest pixel value in image/series.
