+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.2
+1.4.3rc0
