### 2022-03-22 08d7156
Header.__set_axes_from_template(): Ensure new axis length is correct for present data set. Header.__adjust_axis_from_template(): Construct new axis from template.
