### 2022-03-22 7350f71
test_series.test_header_axis: Verify that template axes are adjusted to present data.
