### 2022-03-22 387b7cd
setup.cfg: Specify exact python versions supported.
