+++ setup.cfg
@@ -21,7 +21,10 @@ keywords =
     pynetdicom
     itk
 classifiers =
-        Programming Language :: Python :: 3
+        Programming Language :: Python :: 3.6
+        Programming Language :: Python :: 3.7
+        Programming Language :: Python :: 3.8
+        Programming Language :: Python :: 3.9
         Development Status :: 4 - Beta
         Environment :: Other Environment
         Intended Audience :: Developers
