### 2022-03-22 #4133 @erling6232
Dear all,

Thanks for your effort in reviewing this work. I appreciate it!

There are automated tests implemented as github actions, which run through all unittests (approx 190 tests) on every update. These tests are successful on python 3.6-3.9.

Thanks for point out that community guidelines were missing. A CONTRIBUTING.rst file has been added!
