+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.3rc0
+1.4.3
