### 2022-04-04 1862a1c
ZipfilesystemArchive: Use relative imports for imagedata modules. ZipfilesystemArchive: Updated docstrings. ZipfilesystemArchive.open(): Argument is member, not filehandle. ZipfilesystemArchive.is_file(): Argument is member, not filehandle. ZipfilesystemArchive.to_localfile(): Argument is member, not filehandle.
