+++ src/imagedata/archives/zipfilearchive.py
@@ -1,7 +1,7 @@
-"""Read/Write image files from a zipfile
+"""Read/Write files from a zipfile
 """
 
-# Copyright (c) 2018-2021 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2018-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
 import os
 import os.path
@@ -13,9 +13,8 @@ import urllib.parse
 import logging
 from abc import ABC
 
-import imagedata.archives
-import imagedata.transports
-from imagedata.archives.abstractarchive import AbstractArchive
+from .abstractarchive import AbstractArchive, Member
+from ..transports import Transport
 import zipfile
 
 logger = logging.getLogger(__name__)
@@ -36,37 +35,37 @@ class WriteFileIO(io.FileIO):
     """Local object making sure the new file is written to zip
     archive before closing."""
 
-    def __init__(self, archive, filename, localfile):
+    def __init__(self, archive, member, local_file):
         """Make a WriteFileIO object.
 
         Args:
             archive: ZipFile object
-            filename: path name in zip archive
-            localfile: path name to local, temporary file
+            member: member of the zip archive
+            local_file: local temporary file
         """
-        super(WriteFileIO, self).__init__(localfile.name, mode='wb')
+        super(WriteFileIO, self).__init__(local_file.name, mode='wb')
         self.__archive = archive
-        self.__filename = filename
-        self.__localfile = localfile
+        self.__filename = member.filename
+        self.__local_file = local_file
 
     def close(self):
         """Close file, copy it to archive, then delete local file."""
         logger.debug("ZipfileArchive.WriteFileIO.close:")
         ret = super(WriteFileIO, self).close()
-        self.__localfile.close()
+        self.__local_file.close()
         logger.debug("ZipfileArchive.WriteFileIO.close: zip %s as %s" %
-                     (self.__localfile.name, self.__filename))
-        self.__archive.write(self.__localfile.name, self.__filename)
+                     (self.__local_file.name, self.__filename))
+        self.__archive.write(self.__local_file.name, self.__filename)
         logger.debug("ZipfileArchive.WriteFileIO.close: remove %s" %
-                     self.__localfile.name)
-        os.remove(self.__localfile.name)
+                     self.__local_file.name)
+        os.remove(self.__local_file.name)
         return ret
 
     def __enter__(self):
         """Enter context manager.
         """
         logger.debug("ZipfileArchive.WriteFileIO __enter__: %s %s" %
-                     (self.__filename, self.__localfile.name))
+                     (self.__filename, self.__local_file.name))
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
@@ -83,7 +82,7 @@ class ZipfileArchive(AbstractArchive, ABC):
     name = "zip"
     description = "Read and write image files from a zipfile."
     authors = "Erling Andersen"
-    version = "1.0.0"
+    version = "1.1.0"
     url = "www.helse-bergen.no"
     mimetypes = ['application/zip', 'application/x-zip-compressed']
 
@@ -96,11 +95,11 @@ class ZipfileArchive(AbstractArchive, ABC):
     # self.__tmpdir: Local directory where zip file is unpacked.
     # self.__files: dict of files in the zip archive.
     #   key is path name in the zip archive.
-    #   value is a dict of member info:
-    #     'unpacked': whether the file is unpacked in tmpdir (boolean)
-    #     'name': path name in the zip archive
-    #     'fh': file handle when open, otherwise None
-    #     'localfile': local filename of unpacked file
+    #   value is a Member object:
[…]
