### 2022-04-04 6bbfe70
AbstractPlugin: Use relative imports for imagedata modules. AbstractPlugin.read(): Do not scan files when source['files'] is empty. AbstractPlugin.read(): Use try..finally to ensure files are closed.
