### 2022-04-04 6a312de
DICOMPlugin: Use relative imports for imagedata modules. DICOMPlugin.get_dicom_headers(): No deep scan for filenames when source['files'] is empty.
