+++ src/imagedata/formats/dicomplugin.py
@@ -1,7 +1,7 @@
 """Read/Write DICOM files
 """
 
-# Copyright (c) 2013-2019 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2013-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
 import os
 import logging
@@ -14,10 +14,11 @@ import pydicom.errors
 import pydicom.uid
 from pydicom.datadict import tag_for_keyword
 
-import imagedata.formats
-import imagedata.axis
-from imagedata.formats.abstractplugin import AbstractPlugin
-from imagedata.header import Header
+from ..formats import CannotSort, NotImageError, INPUT_ORDER_FAULTY, input_order_to_dirname_str, SORT_ON_SLICE,\
+    INPUT_ORDER_NONE, INPUT_ORDER_TIME, INPUT_ORDER_B, INPUT_ORDER_FA, INPUT_ORDER_TE
+from ..axis import VariableAxis, UniformLengthAxis
+from .abstractplugin import AbstractPlugin
+from ..header import Header
 
 logger = logging.getLogger(__name__)
 
@@ -247,11 +248,11 @@ class DICOMPlugin(AbstractPlugin):
         # pydicom.config.debug(True)
         try:
             hdr, shape = self.read_headers(sources, input_order, opts, skip_pixels=True)
-        except imagedata.formats.CannotSort:
+        except CannotSort:
             raise
         except Exception as e:
             logger.debug('DICOMPlugin.read: exception\n%s' % e)
-            raise imagedata.formats.NotImageError('{}'.format(e))
+            raise NotImageError('{}'.format(e))
         # pydicom.config.debug(False)
 
         # Look-up first image to determine pixel type
@@ -276,7 +277,7 @@ class DICOMPlugin(AbstractPlugin):
             hdr.color = True
             shape = shape + (im.SamplesPerPixel,)
             hdr.axes.append(
-                imagedata.axis.VariableAxis(
+                VariableAxis(
                     'rgb',
                     ['r', 'g', 'b']
                 )
@@ -489,7 +490,7 @@ class DICOMPlugin(AbstractPlugin):
         #    #import traceback
         #    #traceback.print_exc()
         #    #logger.info("process_member: Could not read {}".format(member_name))
-        #    raise imagedata.formats.NotImageError(
+        #    raise NotImageError(
         #        'Does not look like a DICOM file: {}'.format(sources))
         # except Exception as e:
         #    logger.debug('DICOMPlugin.read_headers: Exception {}'.format(e))
@@ -520,11 +521,10 @@ class DICOMPlugin(AbstractPlugin):
             archive = source['archive']
             scan_files = source['files']
             logger.debug("DICOMPlugin.get_dicom_headers: archive: {}".format(archive))
-            if scan_files is None or len(scan_files) == 0:
-                scan_files = ['*']
+            # if scan_files is None or len(scan_files) == 0:
+            #     scan_files = ['*']
             logger.debug("get_dicom_headers: source: {} {}".format(type(source), source))
             logger.debug("get_dicom_headers: scan_files: {}".format(scan_files))
-            # for member in archive.getmembers(scan_files):
             for path in archive.getnames(scan_files):
                 logger.debug("get_dicom_headers: member: {}".format(path))
                 if os.path.basename(path) == "DICOMDIR":
@@ -542,7 +542,6 @@ class DICOMPlugin(AbstractPlugin):
                         self.process_member(image_dict, archive, path, f, opts, skip_pixels=skip_pixels)
                 except Exception as e:
                     logger.debug('DICOMPlugin.get_dicom_headers: Exception {}'.format(e))
[…]
