### 2022-04-04 27819d4
cmdline: Use relative imports for imagedata modules.
