+++ setup.cfg
@@ -45,7 +45,7 @@ packages = find_namespace:
 python_requires = >=3.6,<3.10
 install_requires =
     importlib-metadata; python_version < '3.8'
-    pydicom>=2.1.0
+    pydicom>=2.1.0, <2.3.0
     pynetdicom>=1.2.0, <2
     itk-io>=5.0
     nibabel
