### 2022-04-04 9ea17e0
At present, imagedata is not compatible with latest pydicom 2.3. Use up to pydicom 2.2 only.
