+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.3
+1.4.4rc0
