### 2022-04-04 ab50a25
Transport.open() returns IO[bytes].
