+++ setup.cfg
@@ -45,6 +45,8 @@ packages = find_namespace:
 python_requires = >=3.6,<3.10
 install_requires =
     importlib-metadata; python_version < '3.8'
+    typing_extensions; python_version < '3.8'
+    future_annotations; python_version < '3.7'
     pydicom>=2.1.0
     pynetdicom>=1.2.0, <2
     itk-io>=5.0
