+++ tox.ini
@@ -4,7 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py36, py37, py38, py39
+envlist = py37, py38, py39
 isolated_build = True
 
 [testenv]
