+++ src/imagedata/archives/filesystemarchive.py
@@ -1,19 +1,32 @@
-"""Read/Write local image files
+"""Read/Write local files
 """
 
-# Copyright (c) 2018 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2018-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
+from __future__ import annotations
+from abc import ABC
+# from typing import IO, Literal, Optional, Type
+from typing import IO, Optional, Type
+# Import Literal from typing_extension when Python < 3.8
+try:
+    from typing_extensions import Literal
+except ImportError:
+    try:
+        from typing import Literal
+    except ImportError:
+        Literal = None
+from types import TracebackType
 import os
 import os.path
 import fnmatch
 import shutil
 import urllib.parse
 import logging
-from abc import ABC
 
-import imagedata.archives
-import imagedata.transports
-from imagedata.archives.abstractarchive import AbstractArchive
+from .abstractarchive import AbstractArchive, Member
+from . import FileAlreadyExistsError
+from imagedata.transports import Transport, RootDoesNotExist, RootIsNotDirectory
+from imagedata.transports.abstracttransport import AbstractTransport
 
 logger = logging.getLogger(__name__)
 
@@ -31,53 +44,40 @@ class NoSuchFile(Exception):
 
 
 class FilesystemArchive(AbstractArchive, ABC):
-    """Read/write local files."""
+    """Read/write local files.
+
+    Args:
+        transport: a Transport instance
+        url (str): URL to filesystem
+        mode (str): filesystem access mode
+        read_directory_only (bool): Whether url should refer to a directory.
+        opts (dict): Options
+
+    Returns:
+        FilesystemArchive instance
+    """
 
     name = "filesystem"
-    description = "Read and write local image files."
+    description = "Read and write local files."
     authors = "Erling Andersen"
-    version = "1.0.0"
+    version = "1.1.0"
     url = "www.helse-bergen.no"
     mimetypes = ['*']  # Disregards MIME types
 
-    # self.__dirname: root directory
-    # self.__filelist: list of absolute filename
+    # Internal data
+    __netloc: str
+    __path: str
+    __transport: AbstractTransport
+    __mode: Literal["r", "w", "x", "a"]  # read or write access
+    __dirname: str
+    __basename: str
 
-    @staticmethod
-    def _get_transport(url, mode, read_directory_only):
-        """Get transport plugin from url.
-
-        If the url addressess a missing file in read mode,
-        access the parent directory.
-        """
-
-        url_tuple = urllib.parse.urlsplit(url, scheme='file')
-        logger.debug('FilesystemArchive._get_transport: scheme: %s, netloc: %s' %
-                     (url_tuple.scheme, url_tuple.path))
-
-        try:
-            _transport = imagedata.transports.Transport(
-                url_tuple.scheme,
-                netloc=url_tuple.netloc,
-                root=url_tuple.path,
-                mode=mode,
-                read_directory_only=read_directory_only)
-        except imagedata.transports.RootDoesNotExist:
-            # Mode='r': location does not exist
-            raise
-        except imagedata.transports.RootIsNotDirectory:
-            # Mode='r': Retry with parent directory
-            parent, _ = os.path.split(url_tuple.path)
-            _transport = imagedata.transports.Transport(
-                url_tuple.scheme,
-                netloc=url_tuple.netloc,
-                root=parent,
-                mode=mode,
-                read_directory_only=read_directory_only)
-        return _transport
-
-    def __init__(self, transport=None, url=None, mode='r',
-                 read_directory_only=True, opts=None):
+    def __init__(self,
+                 transport: Optional[AbstractTransport] = None,
[…]
