+++ src/imagedata/archives/abstractarchive.py
@@ -1,11 +1,16 @@
-"""Abstract class for image archives.
+"""Abstract class for archives.
 
 Defines generic functions.
 """
 
-# Copyright (c) 2018 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2018-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
+from __future__ import annotations
 from abc import ABCMeta, abstractmethod
+from typing import IO, Optional, Type
+from types import TracebackType
+import io
+from imagedata.transports.abstracttransport import AbstractTransport
 
 
 class NoOtherInstance(Exception):
@@ -16,10 +21,28 @@ class WriteMultipleArchives(Exception):
     pass
 
 
+class Member(object):
+    """Class definition for filehandle in imagedata archives.
+    """
+
+    def __init__(self, filename: str,
+                 info: Optional[dict] = None,
+                 fh: Optional[io.FileIO] = None,
+                 local_file: Optional[str] = None):
+        """Initialize the filehandle object."""
+        self.filename = filename
+        if info is None:
+            self.info = {}
+        else:
+            self.info = info
+        self.fh = fh
+        self.local_file = local_file
+
+
 class AbstractArchive(object, metaclass=ABCMeta):
     """Abstract base class definition for imagedata archive plugins.
     Plugins must be a subclass of AbstractPlugin and
-    must define the atttributes set in __init__() and
+    must define the attributes set in __init__() and
     the following methods:
 
     __init__() method
@@ -29,13 +52,18 @@ class AbstractArchive(object, metaclass=ABCMeta):
     open() method
     getmembers() method
     to_localfile() method
-    read() method
-    write() method
+    add_localfile() method
+    writedata() method
+    is_file() method
     """
 
     plugin_type = 'archive'
 
-    def __init__(self, name, description, authors, version, url, _mimetypes):
+    def __init__(self, name: str, description: str, authors: str,
+                 version: str, url: str,
+                 _mimetypes: list):
+        """Initialize the archive object.
+        """
         object.__init__(self)
         self.__name = name
         self.__description = description
@@ -44,15 +72,14 @@ class AbstractArchive(object, metaclass=ABCMeta):
         self.__url = url
         self.__mimetypes = _mimetypes
         self.__transport = None
-        """Return an archive object for filehandle."""
 
     @abstractmethod
-    def use_query(self):
-        """Do the plugin need the ?query part of the url?"""
+    def use_query(self) -> bool:
+        """Does the plugin need the ?query part of the url?"""
         pass
 
     @property
-    def name(self):
+    def name(self) -> str:
         """Plugin name
         
         Single word string describing the image format.
@@ -61,7 +88,7 @@ class AbstractArchive(object, metaclass=ABCMeta):
         return self.__name
 
     @property
-    def description(self):
+    def description(self) -> str:
         """Plugin description
         
         Single line string describing the image format.
@@ -69,7 +96,7 @@ class AbstractArchive(object, metaclass=ABCMeta):
         return self.__description
 
     @property
-    def authors(self):
+    def authors(self) -> str:
         """Plugin authors
         
         Multi-line string naming the author(s) of the plugin.
@@ -77,7 +104,7 @@ class AbstractArchive(object, metaclass=ABCMeta):
         return self.__authors
 
     @property
-    def version(self):
+    def version(self) -> str:
         """Plugin version
         
         String giving the plugin version.
@@ -86,7 +113,7 @@ class AbstractArchive(object, metaclass=ABCMeta):
         return self.__version
 
     @property
-    def url(self):
+    def url(self) -> str:
         """Plugin URL
         
         URL string to the site of the plugin or the author(s).
@@ -94,22 +121,25 @@ class AbstractArchive(object, metaclass=ABCMeta):
[…]

[…]
