+++ src/imagedata/transports/xnattransport.py
@@ -1,18 +1,19 @@
 """Read/write files in xnat database
 """
 
-# Copyright (c) 2021 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2021-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
 import os
 import os.path
 import io
+from typing import IO
 import fnmatch
 import logging
 import shutil
 import tempfile
 import urllib
 import xnat
-from imagedata.transports.abstracttransport import AbstractTransport
+from .abstracttransport import AbstractTransport
 
 logger = logging.getLogger(__name__)
 
@@ -196,7 +197,7 @@ class XnatTransport(AbstractTransport):
         """
         pass
 
-    def open(self, path, mode='r'):
+    def open(self, path, mode='r') -> IO[bytes]:
         """Extract a member from the archive as a file-like object.
         """
         if mode[0] == 'r' and not self.__local:
