+++ src/imagedata/formats/matplugin.py
@@ -1,16 +1,17 @@
 """Read/Write Matlab-compatible MAT files
 """
 
-# Copyright (c) 2013-2018 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2013-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
 import os.path
 import logging
 import numpy as np
 import scipy
 import scipy.io
-import imagedata.formats
-import imagedata.axis
-from imagedata.formats.abstractplugin import AbstractPlugin
+from . import NotImageError, input_order_to_dirname_str, WriteNotImplemented, shape_to_str, sort_on_to_str,\
+    SORT_ON_SLICE
+from ..axis import UniformLengthAxis
+from .abstractplugin import AbstractPlugin
 
 logger = logging.getLogger(__name__)
 
@@ -93,8 +94,8 @@ class MatPlugin(AbstractPlugin):
             logger.debug('matplugin._read_image variable {}'.format(name))
             si = self._reorder_to_dicom(mdict[name])
             logger.info("Data shape _read_image MAT: {} {}".format(si.shape, si.dtype))
-        except imagedata.formats.NotImageError:
-            raise imagedata.formats.NotImageError('{} does not look like a MAT file'.format(f))
+        except NotImageError:
+            raise NotImageError('{} does not look like a MAT file'.format(f))
         return info, si
 
     def _set_tags(self, image_list, hdr, si):
@@ -113,12 +114,12 @@ class MatPlugin(AbstractPlugin):
         hdr.spacing = (1.0, 1.0, 1.0)
 
         axes = list()
-        axes.append(imagedata.axis.UniformLengthAxis(
+        axes.append(UniformLengthAxis(
             'row',
             0,
             si.shape[-2])
         )
-        axes.append(imagedata.axis.UniformLengthAxis(
+        axes.append(UniformLengthAxis(
             'column',
             0,
             si.shape[-1])
@@ -127,15 +128,15 @@ class MatPlugin(AbstractPlugin):
         nt = nz = 1
         if si.ndim > 2:
             nz = si.shape[-3]
-            axes.insert(0, imagedata.axis.UniformLengthAxis(
+            axes.insert(0, UniformLengthAxis(
                 'slice',
                 0,
                 nz)
                         )
         if si.ndim > 3:
             nt = si.shape[-4]
-            axes.insert(0, imagedata.axis.UniformLengthAxis(
-                imagedata.formats.input_order_to_dirname_str(hdr.input_order),
+            axes.insert(0, UniformLengthAxis(
+                input_order_to_dirname_str(hdr.input_order),
                 0,
                 nt)
                         )
@@ -168,7 +169,7 @@ class MatPlugin(AbstractPlugin):
         """
 
         if si.color:
-            raise imagedata.formats.WriteNotImplemented(
+            raise WriteNotImplemented(
                 "Writing color MAT images not implemented.")
 
         logger.debug('MatPlugin.write_3d_numpy: destination {}'.format(destination))
@@ -181,7 +182,7 @@ class MatPlugin(AbstractPlugin):
         self.spacing = si.spacing
         self.tags = si.tags
 
-        logger.info("Data shape write: {}".format(imagedata.formats.shape_to_str(si.shape)))
+        logger.info("Data shape write: {}".format(shape_to_str(si.shape)))
         if si.ndim == 4 and si.shape[0] == 1:
             si.shape = si.shape[1:]
         # if si.ndim == 2:
@@ -192,7 +193,7 @@ class MatPlugin(AbstractPlugin):
[…]
