+++ src/imagedata/__init__.py
@@ -12,9 +12,11 @@ except ModuleNotFoundError:
     from importlib_metadata import version, entry_points
     __version__ = version('imagedata')
 except Exception:
-    import imagedata as _
+    # import imagedata as _
+    from . import __path__ as _path
     from os.path import join
-    with open(join(_.__path__[0], "..", "VERSION.txt"), 'r') as fh:
+    # with open(join(_.__path__[0], "..", "VERSION.txt"), 'r') as fh:
+    with open(join(_path[0], "..", "VERSION.txt"), 'r') as fh:
         __version__ = fh.readline().strip()
 
 __author__ = 'Erling Andersen, Haukeland University Hospital, Bergen, Norway'
