+++ src/imagedata/series.py
@@ -19,11 +19,12 @@ import pydicom.dataset
 import pydicom.datadict
 # import numpy.core._multiarray_umath
 
-import imagedata.axis
-import imagedata.formats
-import imagedata.readdata as readdata
-import imagedata.formats.dicomlib.uid
-from imagedata.header import Header
+from .axis import UniformAxis, UniformLengthAxis, VariableAxis
+from .formats import INPUT_ORDER_NONE, INPUT_ORDER_TIME, INPUT_ORDER_B
+from .formats import input_order_to_dirname_str, shape_to_str, input_order_set, sort_on_set
+from .formats.dicomlib.uid import get_uid_for_storage_class
+from .readdata import read as r_read, write as r_write
+from .header import Header, deepcopy_DicomHeaderDict
 
 logger = logging.getLogger(__name__)
 
@@ -129,14 +130,14 @@ class Series(np.ndarray):
             obj.header.input_order = input_order
             obj.header.input_format = type(data)
             if np.ndim(data) == 0:
-                obj.header.axes = [imagedata.axis.UniformAxis('number', 0, 1)]
+                obj.header.axes = [UniformAxis('number', 0, 1)]
             obj.header.set_default_values(obj.axes if axes is None else axes)
             obj.header.add_template(template)
             obj.header.add_geometry(template, geometry)
             return obj
 
         # Read input, hdr is dict of attributes
-        hdr, si = readdata.read(urls, input_order, opts)
+        hdr, si = r_read(urls, input_order, opts)
 
         obj = np.asarray(si).view(cls)
         assert obj.header, "No Header found in obj.header"
@@ -285,7 +286,7 @@ class Series(np.ndarray):
                 if input_.header is None:
                     logger.warning('Series._unify_headers: new header')
                     header = Header()
-                    header.input_order = imagedata.formats.INPUT_ORDER_NONE
+                    header.input_order = INPUT_ORDER_NONE
                 else:
                     logger.debug('Series._unify_headers: copy header')
                     header = copy.copy(input_.header)
@@ -413,7 +414,7 @@ class Series(np.ndarray):
                     todo.append(('imagePositions', None))  # Wipe existing positions
                     if ipp is not None:
                         todo.append(('imagePositions', ipp))
-                elif axis.name == imagedata.formats.input_order_to_dirname_str(self.input_order):
+                elif axis.name == input_order_to_dirname_str(self.input_order):
                     # Select slice of tags
                     tags = self.__get_tags(spec)
                     todo.append(('tags', tags))
@@ -432,7 +433,7 @@ class Series(np.ndarray):
                 # Must copy the ret object before modifying. Otherwise, ret is a view to self.
                 ret.header = copy.copy(ret.header)
                 if ret.axes[-ret.ndim].name in ['slice', 'row', 'column', 'color']:
-                    ret.input_order = imagedata.formats.INPUT_ORDER_NONE
+                    ret.input_order = INPUT_ORDER_NONE
                 else:
                     raise IndexError('Unexpected axis {} after slicing'.format(ret.axes[0].name))
             _set_geometry(ret, todo)
@@ -496,7 +497,7 @@ class Series(np.ndarray):
             start, stop, step, axis = specs[d]
             if axis.name == 'slice':
                 slice_spec = slice(start, stop, step)
-            elif axis.name == imagedata.formats.input_order_to_dirname_str(self.input_order):
+            elif axis.name == input_order_to_dirname_str(self.input_order):
                 tag_spec = slice(start, stop, step)
         # DicomHeaderDict[slice].tuple(tagvalue, filename, dicomheader)
         hdr = {}
[…]
