+++ src/imagedata/readdata.py
@@ -1,7 +1,7 @@
 """Read/Write image files, calling appropriate transport, archive and format plugins
 """
 
-# Copyright (c) 2013-2019 Erling Andersen, Haukeland University Hospital, Bergen, Norway
+# Copyright (c) 2013-2022 Erling Andersen, Haukeland University Hospital, Bergen, Norway
 
 import os.path
 import logging
@@ -10,10 +10,10 @@ import argparse
 import fnmatch
 import pathlib
 import urllib.parse
-import imagedata.formats
-import imagedata
-import imagedata.transports
-import imagedata.archives
+from .formats import INPUT_ORDER_NONE, input_order_to_str, find_plugin, get_plugins_list
+from .formats import CannotSort, NotImageError, UnknownInputError, WriteNotImplemented
+from .transports import RootIsNotDirectory
+from .archives import find_mimetype_plugin, ArchivePluginNotFound
 
 
 class NoTransportError(Exception):
@@ -48,8 +48,8 @@ def read(urls, order=None, opts=None):
         ValueError: When no sources are given.
         UnknownOptionType: When opts cannot be made into a dict.
         FileNotFoundError: When specified URL cannot be opened.
-        imagedata.formats.UnknownInputError: When the input format could not be determined.
-        imagedata.formats.CannotSort: When input data cannot be sorted.
+        UnknownInputError: When the input format could not be determined.
+        CannotSort: When input data cannot be sorted.
     """
 
     logger.debug("reader.read: urls {}".format(urls))
@@ -74,19 +74,19 @@ def read(urls, order=None, opts=None):
                                                                     opts))
 
     # Let the calling party override a default input order
-    input_order = imagedata.formats.INPUT_ORDER_NONE
+    input_order = INPUT_ORDER_NONE
     if 'input_order' in in_opts:
         input_order = in_opts['input_order']
     if order != 'none':
         input_order = order
-    logger.info("Input order: {}.".format(imagedata.formats.input_order_to_str(input_order)))
+    logger.info("Input order: {}.".format(input_order_to_str(input_order)))
 
     # Pre-fetch DICOM template
     pre_hdr = None
     if 'template' in in_opts and in_opts['template']:
         logger.debug("readdata.read template {}".format(in_opts['template']))
         template_source = _get_sources(in_opts['template'], mode='r')
-        reader = imagedata.formats.find_plugin('dicom')
+        reader = find_plugin('dicom')
         pre_hdr, _ = reader.read_headers(template_source, input_order, in_opts)
 
     # Pre-fetch DICOM geometry
@@ -94,14 +94,14 @@ def read(urls, order=None, opts=None):
     if 'geometry' in in_opts and in_opts['geometry']:
         logger.debug("readdata.read geometry {}".format(in_opts['geometry']))
         geometry_source = _get_sources(in_opts['geometry'], mode='r')
-        reader = imagedata.formats.find_plugin('dicom')
+        reader = find_plugin('dicom')
         geom_hdr, _ = reader.read_headers(geometry_source, input_order, in_opts)
         # if pre_hdr is None:
         #    pre_hdr = {}
         # _add_dicom_geometry(pre_hdr, geom_hdr)
 
     # Call reader plugins in turn to read the image data
-    plugins = sorted_plugins_dicom_first(imagedata.formats.get_plugins_list())
+    plugins = sorted_plugins_dicom_first(get_plugins_list())
     logger.debug("readdata.read plugins length {}".format(len(plugins)))
     for pname, ptype, pclass in plugins:
         logger.debug("%20s (%8s) %s" % (pname, ptype, pclass.description))
@@ -115,9 +115,9 @@ def read(urls, order=None, opts=None):
             hdr.add_template(pre_hdr)
             hdr.add_geometry(pre_hdr, geom_hdr)
             return hdr, si
[…]
