+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.4rc0
+1.4.4rc1
