### 2022-04-06 4d4ecaf
Allow pydicom 2.3.
