### 2022-04-06 06b75a7
DICOMPlugin: Silence encoding warnings from pydicom.
