+++ docs/Tutorial.rst
@@ -132,13 +132,13 @@ You can specify a different image format, e.g.:
 
 .. code-block:: python
 
-    a.write('my_itk_dir', 'formats=['itk'])
+    a.write('my_itk_dir', formats=['itk'])
 
 or even multiple formats, where '%p' is replaced be the format name:
 
 .. code-block:: python
 
-    a.write('my_dirs/%p', 'formats=['nifti', 'mat'])
+    a.write('my_dirs/%p', formats=['nifti', 'mat'])
 
 This will save the data in Nifti format in 'my_dirs/nifti', and
 in Matlab format in 'my_dirs/mat'.
