### 2022-04-25 dddce0c
__make_DicomHeaderDict_from_template(): Catch IndexError exception when template size does not match data size.
