+++ CONTRIBUTING.rst
@@ -40,16 +40,15 @@ Developer Documentation
 
 See `Developer Documentation`_ for details on the structure of the code.
 
-.. _submit:
 
 ---------------------
 Submission Guidelines
 ---------------------
-
-.. _submit-issue:
+.. _submit:
 
 Submitting an Issue
 -------------------
+.. _submit-issue:
 
 Before you submit an issue, please search the issue tracker, maybe an issue
 for your problem already exists and the discussion might inform you of
@@ -77,10 +76,9 @@ are going to close an issue that doesn't have enough info to be reproduced.
 You can file new issues by filling out our
 [new issue form](https://github.com/erling6232/imagedata/issues/new).
 
-.. _submit-pr:
-
 How to contribute
 -----------------
+.. _submit-pr:
 
 The preferred workflow for contributing to imagedata is to fork the
 [main repository](https://github.com/erling6232/imagedata) on
