+++ CONTRIBUTING.rst
@@ -20,19 +20,20 @@ create an issue where you can discuss this before creating a pull request.
 Found a Bug?
 ------------
 If you find a bug in the source code, you can help us by
-[submitting an issue](submit-issue_) to our [GitHub Repository][github_].
+`submitting an issue`_ to our [GitHub Repository][github_].
 Even better, you can [submit a Pull Request](submit-pr_) with a fix.
 
 ------------------
 Missing a Feature?
 ------------------
-You can *request* a new feature by [submitting an issue](submit-issue_) to our GitHub Repository.
+You can *request* a new feature by `submitting an issue`_ to our GitHub Repository.
 If you would like to *implement* a new feature, please submit an issue with a proposal for your work first, to be sure that we can use it. Please consider what kind of change it is:
 
 * For a **Major Feature**, first open an issue and outline your proposal so that it can be
-discussed. This will also allow us to better coordinate our efforts, prevent duplication of work,
-and help you to craft the change so that it is successfully accepted into the project.
-* **Small Features** can be crafted and directly [submitted as a Pull Request](submit-pr_).
+  discussed. This will also allow us to better coordinate our efforts, prevent duplication of work,
+  and help you to craft the change so that it is successfully accepted into the project.
+
+* **Small Features** can be crafted and directly `submitted as a Pull Request <submit-pr_>`_.
 
 -----------------------
 Developer Documentation
@@ -48,7 +49,7 @@ Submission Guidelines
 
 Submitting an Issue
 -------------------
-.. _submit-issue:
+.. _submitting an issue:
 
 Before you submit an issue, please search the issue tracker, maybe an issue
 for your problem already exists and the discussion might inform you of
