+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.4
+1.4.5dev0
