+++ CONTRIBUTING.rst
@@ -20,13 +20,13 @@ create an issue where you can discuss this before creating a pull request.
 Found a Bug?
 ------------
 If you find a bug in the source code, you can help us by
-`submitting an issue <submit-issue_>`_ to our `GitHub Repository <github_>`_.
+`submitting an issue <Submitting an Issue_>`_ to our `GitHub Repository <github_>`_.
 Even better, you can `submit a Pull Request <How to contribute_>`_ with a fix.
 
 ------------------
 Missing a Feature?
 ------------------
-You can *request* a new feature by `submitting an issue <submit-issue_>`_ to our GitHub Repository.
+You can *request* a new feature by `submitting an issue <Submitting an Issue_>`_ to our GitHub Repository.
 If you would like to *implement* a new feature, please submit an issue
 with a proposal for your work first, to be sure that we can use it.
 Please consider what kind of change it is:
@@ -35,22 +35,20 @@ Please consider what kind of change it is:
   discussed. This will also allow us to better coordinate our efforts, prevent duplication of work,
   and help you to craft the change so that it is successfully accepted into the project.
 
-* **Small Features** can be crafted and directly `submitted as a Pull Request <How to contribute>`_.
+* **Small Features** can be crafted and directly `submitted as a Pull Request <How to contribute_>`_.
 
 -----------------------
 Developer Documentation
 -----------------------
 
-See `Developer Documentation`_ for details on the structure of the code.
+See the `Developer Documentation`_ for details on the structure of the code.
 
 -------------------------------
-Submission Guidelines _`submit`
+_`Submission Guidelines`
 -------------------------------
 
-.. _submit-issue:
-
-Submitting an Issue
--------------------
+_`Submitting an Issue`
+----------------------
 
 Before you submit an issue, please search the issue tracker, maybe an issue
 for your problem already exists and the discussion might inform you of
@@ -76,19 +74,19 @@ complete, and verifiable example, so if we don't hear back from you we
 are going to close an issue that doesn't have enough info to be reproduced.
 
 You can file new issues by filling out our
-[new issue form](https://github.com/erling6232/imagedata/issues/new).
+`new issue form`_.
 
 _`How to contribute`
 --------------------
 
 The preferred workflow for contributing to imagedata is to fork the
-[main repository](https://github.com/erling6232/imagedata) on
+`main repository <github_>`_ on
 GitHub, clone, and develop on a branch. Steps:
 
-1. Fork the [project repository](https://github.com/erling6232/imagedata)
+1. Fork the `project repository <github_>`_
    by clicking on the 'Fork' button near the top right of the page. This creates
    a copy of the code under your GitHub user account. For more details on
-   how to fork a repository see [this guide](https://help.github.com/articles/fork-a-repo/).
+   how to fork a repository see `this guide  <fork-a-repo_>`_.
 
 2. Clone your fork of the imagedata repo from your GitHub account to your local disk:
 
@@ -116,7 +114,7 @@ GitHub, clone, and develop on a branch. Steps:
    squashed into one commit with all commit messages combined. The commit
    messages can be edited during the merge, but it helps if they are clearly
    and briefly showing what has been done in the commit. Check out the 
-   [seven commonly accepted rules](https://www.theserverside.com/video/Follow-these-git-commit-message-guidelines)
+   `seven commonly accepted rules <seven-commonly-accepted-rules_>`_
    for commit messages.
    
 6. To record your changes in Git, push the changes to your GitHub
@@ -126,11 +124,11 @@ GitHub, clone, and develop on a branch. Steps:
 
[…]
