+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.5dev0
+1.4.5rc0
