+++ VERSION.txt
@@ -1,1 +1,1 @@
-1.4.5rc0
+1.4.5
