### 2022-04-29 e473cd3
Allow for DICOM objects without pixel array.
