### 2022-04-29 25142ec
Allow for DICOM objects without pixel array. DICOMPlugin.write_object(): New method to write non-pixel objects.
