### 2022-04-29 376f7e1
Catch RuntimeError from ITK reader. Could indicate an input file which ITK does not recognize.
