### 2022-04-29 0fa05b8
readdata.write(): Allow for writing objects with no pixel data.
