### 2022-04-29 8ce2af0
Header.set_default_values(): Set additional DICOM attributes.
