### 2022-05-13 #4133 @galessiorob
Took me a few tries to figure out the case in the handle to remove...

@erling6232 this is so well documented! Thanks for all the work and patience.

The next thing we need from you is to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

@hsang @mwegrzyn Thank you so much for lending your expertise and reviewing this paper 🚀
